update Dict codes, and pass the test of Dict module; StaticDict, DynamicDict and Table is available now.

+++ include/CBL.h
@@ -25,10 +25,9 @@
 #ifndef _CBL_H_
 #define _CBL_H_
 
-#include "config.h"
 #include "CBL_Basic.h"
-#include "CBL_DateTime.h"
 #include "CBL_String.h"
+#include "CBL_DateTime.h"
 #include "CBL_Log.h"
 #include "CBL_Vector.h"
 #include "CBL_Array.h"
@@ -43,6 +42,6 @@
  * @param type typecode defined in `CBL_Basic.h` header file
  * @return bytes of type
  */
-size_t CBL_typecode2size(UInt8 type);
+size_t CBL_typecode2size(Int type);
 
 #endif // _CBL_H_
